requirements.txt
# 投资助手依赖
akshare>=1.12.0      # A股数据
yfinance>=0.2.0      # 美股数据
pyyaml>=6.0          # 配置文件解析
pandas>=2.0.0        # 数据处理
